Break down back office layout
