+++ src/app/(main)/layout.tsx
@@ -1,1 +1,1 @@
-export { Layout as default } from '@/components/Layout';
+export { Layout as default } from '@/ui/main/Layout';
