+++ src/app/layout.tsx
@@ -26,7 +26,7 @@ export default function RootLayout({
     <html
       lang="en"
       className={cn(
-        'min-h-screen bg-background font-sans antialiased',
+        'h-full bg-background font-sans antialiased',
         fontSans.variable
       )}
       suppressHydrationWarning
