+++ src/ui/backoffice/MobileNav.tsx
@@ -1,0 +1,70 @@
+'use client';
+
+import { Disclosure } from '@headlessui/react';
+import { BellIcon, XMarkIcon, Bars3Icon } from '@heroicons/react/20/solid';
+import NavLinks from '@/ui/backoffice/NavLinks';
+import UserNavLinks from './UserNavLinks';
+
+const user = {
+  name: 'Tom Cook',
+  email: 'tom@example.com',
+  imageUrl:
+    'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80'
+};
+
+export default function MobileNav() {
+  return (
+    <>
+      <Disclosure.Panel className="border-b border-gray-700 md:hidden">
+        <div className="space-y-1 px-2 py-3 sm:px-3">
+          <NavLinks mobile={true} />
+        </div>
+        <div className="border-t border-gray-700 pb-3 pt-4">
+          <div className="flex items-center px-5">
+            <div className="flex-shrink-0">
+              <img
+                className="h-10 w-10 rounded-full"
+                src={user.imageUrl}
+                alt=""
+              />
+            </div>
+            <div className="ml-3">
+              <div className="text-base font-medium leading-none text-white">
+                {user.name}
+              </div>
+              <div className="text-sm font-medium leading-none text-gray-400">
+                {user.email}
+              </div>
+            </div>
+            <button
+              type="button"
+              className="relative ml-auto flex-shrink-0 rounded-full bg-gray-800 p-1 text-gray-400 hover:text-white focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800"
+            >
+              <span className="absolute -inset-1.5" />
+              <span className="sr-only">View notifications</span>
+              <BellIcon className="h-6 w-6" aria-hidden="true" />
+            </button>
+          </div>
+
+          <UserNavLinks mobile={true} />
+        </div>
+      </Disclosure.Panel>
+    </>
+  );
+}
+
+export function MobileNavButton({ open }: { open: boolean }) {
+  return (
+    <div className="-mr-2 flex md:hidden">
+      <Disclosure.Button className="relative inline-flex items-center justify-center rounded-md bg-gray-800 p-2 text-gray-400 hover:bg-gray-700 hover:text-white focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800">
+        <span className="absolute -inset-0.5" />
+        <span className="sr-only">Open main menu</span>
+        {open ? (
+          <XMarkIcon className="block h-6 w-6" aria-hidden="true" />
+        ) : (
+          <Bars3Icon className="block h-6 w-6" aria-hidden="true" />
+        )}
+      </Disclosure.Button>
+    </div>
+  );
+}
