+++ src/ui/backoffice/UserNavLinks.tsx
@@ -1,0 +1,61 @@
+'use client';
+
+import { Disclosure, Menu, Transition } from '@headlessui/react';
+import clsx from 'clsx';
+import Link from 'next/link';
+import { Fragment } from 'react';
+
+const userNavigation = [
+  //   { name: 'Your Profile', href: '#' },
+  //   { name: 'Settings', href: '#' },
+  { name: 'Sign out', href: '#' }
+];
+
+export default function UserNavLinks({ mobile }: { mobile?: boolean }) {
+  if (mobile) {
+    return (
+      <div className="mt-3 space-y-1 px-2">
+        {userNavigation.map((item) => (
+          <Disclosure.Button
+            key={item.name}
+            as="a"
+            href={item.href}
+            className="block rounded-md px-3 py-2 text-base font-medium text-gray-400 hover:bg-gray-700 hover:text-white"
+          >
+            {item.name}
+          </Disclosure.Button>
+        ))}
+      </div>
+    );
+  } else {
+    return (
+      <Transition
+        as={Fragment}
+        enter="transition ease-out duration-100"
+        enterFrom="transform opacity-0 scale-95"
+        enterTo="transform opacity-100 scale-100"
+        leave="transition ease-in duration-75"
+        leaveFrom="transform opacity-100 scale-100"
+        leaveTo="transform opacity-0 scale-95"
+      >
+        <Menu.Items className="absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+          {userNavigation.map((item) => (
+            <Menu.Item key={item.name}>
+              {({ active }) => (
+                <Link
+                  href={item.href}
+                  className={clsx(
+                    active ? 'bg-gray-100' : '',
+                    'block px-4 py-2 text-sm text-gray-700'
+                  )}
+                >
+                  {item.name}
+                </Link>
+              )}
+            </Menu.Item>
+          ))}
+        </Menu.Items>
+      </Transition>
+    );
+  }
+}
