+++ src/ui/main/Hero.tsx
@@ -1,4 +1,4 @@
-import { Container } from '@/components/Container';
+import { Container } from '@/ui/main/Container';
 
 export function Hero() {
   return (
