+++ src/ui/main/Layout.tsx
@@ -1,5 +1,5 @@
-import { Footer } from '@/components/Footer';
-import { Header } from '@/components/Header';
+import { Footer } from '@/ui/main/Footer';
+import { Header } from '@/ui/main/Header';
 
 export function Layout({ children }: { children: React.ReactNode }) {
   return (
